Rename callSyncifying to callPromising, add callWithOptions (#4608)
Since we renamed `syncify` the method name `callSyncifying` doesn't make much
sense anymore. It is implemented in terms of a so-called `promising` wasm call,
so `callPromising` lines up with that. Not a perfect name, but it is rarely used so it
will do until we come up with something better.

Added a `callWithOptions` function. We now have three boolean parameters for
a Js-to-Python call:
1. kwargs
2. promising
3. relaxed

So we'd need 8 functions to cover all combinations of these. Currently we only
have 6 of these. Rather than adding the two remaining combinations which will
have annoying names, I added `callWithOptions` which takes an options argument as
the first argument. Despite the fact that options usually go as the last
argument, I think it makes sense to use the first argument for this so that all
remaining args are passed on to Python.