packages: Add lakers-python (#4763)
This adds Lakers (`lakers-python` on PyPI), the Python wrappers for the Rust
implementation of EDHOC (RFC9528). EDHOC is a compact key exchange protocol,
primarily designed for embedded systems.

As a compiled package it can not use the pure Python wheels published through
crates.io. It is useful in pyodide because aiocoap has [good support for pyodide
and Jupyter](https://aiocoap.readthedocs.io/en/latest/pyodide.html) for
end-to-end encryption between a browser console (or application) and an embedded
device (connected through untrusted proxies on the Internet or through local
Bluetooth connections). I expect to add EDHOC support to aiocoap through lakers.