chore: use hatchling instead of setuptools (#4770)
This swaps out the build backend of pyodide-build for hatchling, which is
simpler and faster.

The difference in the SDists: This removes the zip directories entries,
setup.cfg, and .egg-info directory and contents. It includes the .gitignore,
since hatchling uses that to decide what to include by default.

The difference in the wheel built from the SDist: The useless top_level.txt is
gone and the license is in the licenses directory.