Fix dynamic library loading in cli runner (#4871)
All dependent libs should be loaded as globals. We should try  to use this to
simplify some of the logic in dynload.ts. Most likely what needs to happen is we
need to inspect the wheel at build time to determine which libraries are
"leaves" that are not dependencies of anything else and add this list to the
package metadata.

This should be removed eventually, because it is a bug somewhere else in the
dynamic loader that dependents need to be global.