Add CasADi (#4936)
This PR aims to add CasADi, a C++ library with SWIG-generated bindings to Python. CasADi is a
framework for Python and MATLAB (and Octave) which provides tools for nonlinear optimization and
algorithmic differentiation, and offers interfaces to several solver suites, which are not required
but can be turned on via CMake flags. Just the SUNDIALS solver suite is required to build, and the
rest of the interfaces are optional (they are not enabled because of the high build times).