Snapshots: Add support for custom serializer
The user may pass a serializer function to `makeSnapshot`. It will be called
with JavaScript objects that we don't know how to snapshot. It must return an
object that we can `JSON.stringify`. When restoring the snapshot, the user must
pass a corresponding deserializer that will take the object returned by the
serializer and create the original JSON object.