Update stack switching to support new JSPI API (#4982)
The JSPI API changed. Currently v8 implements both the old and the new version.
We want to keep support for the old version for a while so we can use stack
switching in node. This supports both.

I also simplified the implementation by quite a lot somewhat by accident in the
process of trying to work out how to support the new API.

Resolves #4883.