Add tree-sitter-java, tree-sitter-go, tree-sitter-python
This is a follow-on to PR #5099 which added the dependency
of tree-sitter to the packages.

This change adds tree-sitter-java, tree-sitter-go, tree-sitter-python
packages and respective unit tests. Java, Go, and Python are
popular languages, so its useful to add these grammars so a user
can parse source code in those languages.

Signed-off-by: Eric Brown <eric_wade_brown@yahoo.com>