Testing for pyodide/pyodide-build#60
See upstream PR:
https://github.com/pyodide/pyodide-build/pull/60

Resolves pyodide/pyodide#5187. In that issue, the build requires a package
called `cmake-build-extension` but because we have `cmake` in the list of
requirements to avoid we drop it and it breaks the build.

This logic was introduced in pyodide/pyodide#2272 and
did not come up in the review. Surprisingly this is the first problem it has
caused. Let's see if this change breaks anything in Pyodide CI.