Allow calling registerJsModule with a non extensible JS object
Currently we can't use `registerJsModule` with an ES6 module:
```js
registerJsModule("jsfs", await import("fs"))
```
because we try to add a `__all__` property to the JS object but
the ES6 module is non extensible. Instead of adding a `__all__`,
teach `from mod import *` about how to get the properties of a JS
module.