lastModifiedDate fix for issue/5425 (#5426)
This is a fix for issue #5425

File API File.lastModifiedDate property is deprecated by Firefox and Safari, it
is better to use new Date(file.lastModified) instead of file.lastModifiedDate in
nativefs.ts for Firefox and Safari compatibility.

Most likely it'll also fix issue #4057, but unable to test on Safari to confirm.