Fix `thisArg` when making js calls while taking snapshot
If we pass `_makeSnapshot: true` to `loadPyodide`, it wraps `globalThis`
in this `makeGlobalsProxy` function so we can serialize top level imports.
This breaks functions like `fetch` which in Chrome raises:
```
Failed to execute 'fetch' on 'Window': Illegal invocation
```
In order to fix this, before calling a function we should unwrap the target.