Add eager_converter option to `to_js` which takes precedence over default behavior
The `default_converter` argument is called after the builtin conversions so it
cannot be used to modify the normal behavior. dom96 needed a way to block converting
tuples. This adds a way to do this.

One API question is whether we want to allow the default behavior by calling `convert()`
or by returning a special value. I have implemented it via `convert()` which leads to
some slightly messy special case handling. It would be possible to do a fallback when `None`
is returned, but what if the actual intention was to convert to `undefined`? Other special
values would be hard to come by either in Python or in JS or both. So I think this is
fine.