Fix: callPromising should not treat async error as fatal (#5678)
There is a problem here with Promise flattening: if the return value
of the function is a thenable, we accidentally await both the promise
from stack switching and the enclosed thenable. This raises a user error
out and we catch it and treat it as a fatal error.

To avoid this problem, we wrap the result in a one element list to prevent
the promises from being flattened.