Improve the error message when a dynamic library is not found
Prior to this, if we dlopen a dynamic library a.so which depends on
b.so and we find a.so but not b.so, the error message looks like:
```
Could not load dynamic lib: a.so
Error: Didn't expect to load any more file_packager files!
```
After this change:
```
Could not load dynamic lib: a.so
Error: Failed to find dynamic library b.so
```
So the error message will mention `b.so` now.