Stop using JavaScript null as an error value (#5719)
We need JS null to represent actual null and not an error. Instead, we use wasm-gc except on safari iOS
where wasm-gc is broken or on old runtimes where wasm-gc support is missing. If we can't use wasm-gc,
we use a JS object and a JS trampoline. There is maybe a 3-4% performance hit associated with
this fallback compared to before this PR. Using wasm-gc to signal errors leads to no measured performance
change.