DOCS Improvements to JS API docs (#5807)
I renamed the `Globals` section to `Top Level Exports` and added a comment
explaining that `loadPyodide()` is placed on `globalThis` by `pyodide.js` and
the other exports are left inaccessible.

To make xrefs to objects in this section, it's now necessary to say
`exports.blah` instead of `globalThis.blah`. I suppressed the `exports.` /
`globalThis.` prefix that used to appear on objects in this section. I
suppressed the prefixes for interfaces and type aliases because they cannot be
accessed like that.

I added `PyodideAPI` as a type export in `pyodide.d.ts` and added a
documentation entry for it. I moved `TypedArray` to `pyodide.ffi` in the docs
(the types say that `import {TypedArray} from "pyodide";` is deprecated and it
should be imported from `pyodide.ffi` instead).