Load packages without using Python
If we want to execute .pth files in packages, we need to load the
packages before we initialize the Python interpreter. We also need
this for taking memory snapshots with packages.

This switches to using zip.js to unpack the zip archives. We still
install data files and unpack .tar archives with Python. To handle
this, when preloading packages we pass in an empty
`_postBootstrapPromises` list and push tasks we want deferred to
after interpreter startup into this list.

Installing data files just requires Python to get `sys.prefix`.
Unpacking tar files would require us to also get a JavaScript tar
parser. Probably it would be better to convert tests packages to
the zip format.