PERF Improve performance of `gc_register_pyproxy` (#5975)
According to profiles, 60% of the time spent creating a PyProxy is spent in
`gc_register_proxy`. 63% of the time in `gc_register_proxy` is spent in
`finalizationRegistry.register()` which we can't improve. However, the remaining
37% of the time is spent in this `Object.assign()` call. I'm not sure why it's
so expensive, but this new code is 100 times faster than `Object.assign()`. It
reduces the time to create a PyProxy by over 20%.